Correction - 516 Beijing 2008 DIV 1

+++ src/com/topcoder/room516/beijing2008/div1_250/WaterEqualizer.java
@@ -3,13 +3,13 @@ package com.topcoder.room516.beijing2008.div1_250;
 public class WaterEqualizer {
 
 	public double minTime(int[] waterVolumes) {
-		long allConsist = 0;
+		double allConsist = 0;
 		for (int waterVolum : waterVolumes) {
 			allConsist += waterVolum;
 		}
-		long averageConsist = allConsist / waterVolumes.length;
+		double averageConsist = allConsist / waterVolumes.length;
 
-		long displacedWater = 0;
+		double displacedWater = 0;
 		for (int waterVolum : waterVolumes) {
 			displacedWater += Math.abs(averageConsist - waterVolum);
 		}
